refactor: modify sign-in and sign-up logic
